Replace hash-derived runtime ids with allocator-issued Longs
Map-key conflict identity was hashed from (mapId, key).toString while the
value store keys slots by equals/hashCode — two different equivalence
relations. Equal keys with different renderings (BigDecimal("1.0") vs
"1.00") got two footprint ids for one slot, so the scheduler missed the
conflict and concurrent read-modify-writes lost updates; distinct keys
with one rendering collapsed to one id, falsely serializing unrelated
transactions and clobbering same-transaction log entries. Separately,
every runtime id narrowed a name-UUID to a 32-bit Int, reintroducing the
birthday-collision exposure the UUID step's own comment said it existed
to prevent.

Ids are now Longs issued by the one global allocator: entities use their
TxnVarId directly, and each TxnVarMap issues key existential ids from the
same counter through a per-map registry keyed by the key's own equality.
Raw values are globally unique by construction — required because
IdFootprint compares them parent-stripped. Registry entries are never
evicted (a parked transaction's footprint holds the id); the scaladoc and
CHANGELOG document that along with the key-retention cost.

CommitLockOrderSpec's H2 guard relied on hash coin-flips for opposite
lock orders; under allocation order every transaction would order
m1-before-m2 and a revert could never deadlock. A deterministic pre-touch
phase restores opposed orders — re-verified red against a locally
reverted sort — and KeyIdentitySpec pins the first-touch ordering
property it relies on, both divergence directions, allocate-race
convergence, delete/reinsert stability, and entity/existential
non-collision.

tlBaseVersion moves to 0.15: the Int-to-Long widening changes erased
signatures against the published 0.14.0 and the id semantics change
observably. Stale hash-era prose corrected across TxnLogContext, the
TLA+ specs, specs/README.md, and benchmarks/README.md.

+++ src/main/scala/bengal/stm/model/runtime/IdFootprint.scala
@@ -70,11 +70,11 @@ private[stm] case class IdFootprint(
   private[stm] lazy val combinedIds: Set[TxnVarRuntimeId] =
     readIds ++ updatedIds
 
-  private[stm] lazy val combinedRawIds: Set[Int] = combinedIds.map(_.value)
+  private[stm] lazy val combinedRawIds: Set[Long] = combinedIds.map(_.value)
 
-  private[stm] lazy val updateRawIds: Set[Int] = updatedIds.map(_.value)
+  private[stm] lazy val updateRawIds: Set[Long] = updatedIds.map(_.value)
 
-  private[stm] lazy val readRawIds: Set[Int] = readIds.map(_.value)
+  private[stm] lazy val readRawIds: Set[Long] = readIds.map(_.value)
 
   private[stm] def addReadId(id: TxnVarRuntimeId): IdFootprint =
     this.copy(readIds = readIds + id)
